+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Version 0.2.20 - Not released yet
 * Added method: HelperDate::renderIso8601Date()
 * Added method: HelperDate::toZendDate()
 * Added method: HelperServerClient::addToSessionArray()
+* Added method: HelperServerClient::getRequestScheme()
 * Added ZF1 reflection method: getRequestParamsOfControllerAction()
 * Added ZF1 reflection method: getSourceCodeOfControllerAction()
 * Added ZF1 reflection method: getPhpDocDescriptionOfControllerAction()